pipeline/requirements.txt
apache-beam[gcp]==2.9.0
google-cloud-storage==2.18.2
transforms